Define default env. vars only if not already def'd
The environment variables $KERNALEMU, $CBM6502ASM and $ASSEMBLER64 are
only defined with defaults in build.sh if they are not already defined
yet.

This allows you to set them with

export KERNELEMU=...
export CBM6502ASM=...
export ASSEMBLER64=...

before calling the script, so you do not have to modify the script.